fix: move parsing logic to service
